feat: split lsp module to completion and definition
